Update Zloader.py (#314)
Updated logic on how fields are parsed out of the base config. The RSA key was being parsed in a way that included additional flags set in the config so it has been stripped out for now. The other fields are stored at different offsets that we can reliably check.