add helper function and bloomfilter creation script for network_dga_fraunhofer signature
This signature detects the likely usage of Domain Generation algorithms by
matching all dns requests of a sandbox run against a pre-built bloomfilter.
The bloomfilter contains the active DGA domains from today, tomorrow and
yesterday and is based on the intelligence of the Fraunhofer DGArchive.

The check against the bloomfilter is really quick. Only if we have hits
we'll proceed and match the DGA domain to its DGA/malware family by
performing a dict lookup that is preloaded in memory.

The bloomfilter and the lookup dict are created by the create_bloom.py
script of the CAPEv2 repository and placed within the /data folder of
the community repository. The update process can be performed
manually via running the script on the local machine before
signature push/pull or via cronjob task.

* Bloomfilter library: https://github.com/DCSO/flor
* Fraunhofer DGArchive: https://dgarchive.caad.fkie.fraunhofer.de/welcome/