Precisions on agent placement on Windows machines
I see many people online having issues with the agent because they do not know that the user startup folder will not work. Here is my attempt at clarifying that. Feel free to alter if needed, but this absolutely needs to be in the documentation. I kept the end of the paragraph as-is but I think this should be further improved. Newbie users will have trouble understanding what they have to do with an "etc".