Allow config dirs to be provided that override the default ones.
Add a -c command-line option to cuckoo.py that can be provided multiple
times. This allows the *.conf files in this repo to be untouched by a
user. Instead the values that need to be overridden can be put in to one
or more other directories, whose path is passed via the -c option. The
values in those files will take precedence over the ones in the conf
dir.

In addition, for Config() objects, only read the files once, so that if,
say Config("processing") is created in multiple places, they'll both
refer to the same object and the conf files would only be read once.
Those files are also only read from the first time a section from that
config is referenced. This is because we so often create those Config
objects at module level scope and they could get imported before the
config_dirs have been set. So now, an initialize() classmethod needs to
be called on the Config class before any of those sections are
referenced.

A new AnalysisConfig() class was created to separate the distinction
between these types of "singleton" Config objects that can read from
multiple directories versus ones that just read from an analysis.conf
(or some other single file).