Allow config dirs to be provided that extend the default ones. (#697)
* Allow config dirs to be provided that override the default ones.

Add a -c command-line option to cuckoo.py that can be provided multiple
times. This allows the *.conf files in this repo to be untouched by a
user. Instead the values that need to be overridden can be put in to one
or more other directories, whose path is passed via the -c option. The
values in those files will take precedence over the ones in the conf
dir.

In addition, for Config() objects, only read the files once, so that if,
say Config("processing") is created in multiple places, they'll both
refer to the same object and the conf files would only be read once.
Those files are also only read from the first time a section from that
config is referenced. This is because we so often create those Config
objects at module level scope and they could get imported before the
config_dirs have been set. So now, an initialize() classmethod needs to
be called on the Config class before any of those sections are
referenced.

A new AnalysisConfig() class was created to separate the distinction
between these types of "singleton" Config objects that can read from
multiple directories versus ones that just read from an analysis.conf
(or some other single file).

* Remove use of __getattr__ and ajust some imports.

In cuckoo.py:
- Avoid flake8 complaint about Scheduler not being used in cuckoo_init()
  by just having it imported in cuckoo_main, but report on a cuckoo
  dependency error there, if necessary.
- Pull the import of lib.cuckoo.common.exceptions out of the 'try' block
  since 1) it doesn't itself import any other modules and 2) one of the
  exception types that gets imported from it is used in the 'except' of
  the 'try' block. This could result in a NameError since it might not
  be defined.

In config.py:
- Since we now make sure to not import anything that creates a Config()
  object at the module level until after Config.initialize() is called,
  we don't need to have the lazy loading of the config files in
  `__getattr__`--we can just load them in `__init__`.

Co-authored-by: Tommy Beadle