Work around a bug in pip when doing a fresh install.
Steps to reproduce bug:
 - Create a fresh virtualenv
 - pip install -r requirements.txt
This results in:
```
ERROR: In --require-hashes mode, all requirements must have their
versions pinned with ==. These do not:
    requests from
https://files.pythonhosted.org/packages/2d/61/08076519c80041bc0ffa1a8af0cbd3bf3e2b62af10435d269a9d0f40564d/requests-2.27.1-py2.py3-none-any.whl#sha256=f22fa1e554c9ddfd16e6e41ac79759e17be9e492b3587efa038054674760e72d
(from django-allauth==0.45.0->-r requirements.txt (line 101))
```
This is because of the bug in pip outlined in
https://github.com/pypa/pip/issues/9644.
django-allauth requires `requests` but we only had
`requests[security,socks]` in our requirements.txt and pip didn't see
that as satisfying the `requests` requirement, so it pulls down the
latest version of `requests`, which we don't have hashes for.

This just copies the lines for `requests` without the extras in order to
satisfy that dependency.