Fix logic for finding if a detection was already seen.
The line `if not detection in block["details"]:` would never evaluate to
True, because block["details"] contains references to dicts created in
previous invocations of this function--that statement does not actually
compare the contents of the dicts themselves, but only that they're the
same object.

Also, bail out of the loop when we found the family and use python's
for/else construct instead of tracking `found` ourselves.