Fix bug in CAPE processing.
Some samples may match multiple yara rules that have a cape_type
defined. If the first one has a config extracted but the last one
doesn't, then, previously, self.cape["configs"] would end up having a
an entry with the top-level key of the latter cape_type and its value
would be a dict whose key is the first cape_type and its value would be
the extracted config. This fixes that problem by making sure to add each
config to self.cape["configs"] when it is found instead of waiting until
the very end. This is because the 'cape_type' variable can change as it
goes through iterations of the loop.

I factored the code a little bit to try to make it more readable.

This also fixes a problem where not all cape_names would get added as
family detections. Again, this was because we were only using the last
cape_name that was found.