Fixing bug caused by 6bca51a2c75f9085aa6891c5f338c6804b61533e
That commit removes the distributed section from reporting.conf and moves it to distributed.conf.
Line 528 in web/submission/views.py relied on the distributed section being in the reporting config via the config object,
thus breaking the submission page and requiring this to be updated to the newly created distribuited.conf to fix it.
This commit is meant as a quick patch for the issue caused in the previously referenced commit.