Allow a machine to be marked as "reserved" so that it can only be used if specifically requested.
It a machine has a config that includes "reserved = yes" in the
<machinery>.conf file, then that machine will only be used for a task if
that machine was selected specifically for that task by the user. This
could be useful if you have, say, a new VM that you want to be able to
try out but you don't want it in the normal pool.