Declare pool in process.py
The variable pool is referenced outside of its scope in the finally section of the function autoprocess. An unhandled exception from memory_limit() of type OSError can be thrown, which happens before pool is declared originally. This causes the finally clause to be executed, only pool does not exist yet.