Fix ISO detection without extension
When an ISO is submitted without an extension in the file name, CAPE tries to process it with a package called "iso", but there is no such package. This seems to be caused by a bug in the function checking for archive types which compares the type to a tuple instead of checking if the type is in the tuple. 

You can see an example in this submission:
https://capesandbox.com/analysis/397799/