Two small bug fixes (#1600)
* Declare pool in process.py

The variable pool is referenced outside of its scope in the finally section of the function autoprocess. An unhandled exception from memory_limit() of type OSError can be thrown, which happens before pool is declared originally. This causes the finally clause to be executed, only pool does not exist yet.

* Fix ISO detection without extension

When an ISO is submitted without an extension in the file name, CAPE tries to process it with a package called "iso", but there is no such package. This seems to be caused by a bug in the function checking for archive types which compares the type to a tuple instead of checking if the type is in the tuple. 

You can see an example in this submission:
https://capesandbox.com/analysis/397799/

* Update process.py