Use sudo to invoke tcpdump, if possible.
Rather than having to worry about the capabilities set on tcpdump and
maintaining it whenever tcpdump gets upgraded, just use sudo to invoke
it. This retains backwards compatibility, so it will fall back to check
if tcpdump is suid if it can't be run via sudo.
Update the installer script to add a file to /etc/sudoers.d to configure
it.