Ensure that when a timeout is reached, the SIGTERM is passed to the grandchild process.
When running extraction tools, we use a pebble.ProcessPool. Each tool is
run by calling a function in a fork of the parent process. When a
configured timeout is reached, that process is sent a SIGTERM. Pebble
allows 3 seconds for the process to finish before then sending a SIGKILL
if it's still alive.

If the forked process creates another child, such as by the use of
subprocess.check_output, it would not get the SIGTERM signal. This
change makes sure that that subprocess also gets the signal to
terminate.