Set route to empty string when not specified
When a route is not specified in an API request to create an analysis,
the route would be set to None, eventually resulting in the string
'false', which would then result in no route to be specified, instead of
the default specified in the configuration settings.

Setting it to an empty string if not specified resolves this issue.
Setting to the string 'none' is still possible to explicitly disable
routing.