Don't store an entire text file in "strings".
For text files, the first X bytes (as defined in
processing.conf:CAPE->buffer) are stored in the "data" key of the data
dictionary. We don't need to also store that data in "strings". In fact,
for large text files, previously we could be storing the entire file in
"strings". This could result in the document to be inserted in to
MongoDB being >16MB, which MongoDB doesn't allow. Don't display the
"strings" button in the UI in this case.