feat: better segregation of windows and linux signatures
Previous implementation of segregation defaulted all non-os-labelled signatures to windows. This feature defaulted important static sigantures to only be run on windows analyses. Segregation of signatures to the respective folders, windows, linux and all, allows static-type signatures to be run on all platforms and platform specific running of signaturization. I believe this would also help to improve the time for running signatures.