Make Database creation more explicit and don't initialize it during module import.
Calling Database() just returns a proxy to a real database object that
may or may not exist when the Database object is created. That's fine as
long as no attributes/methods are accessed until the init_database
function has been called, which must be called only once during the
program's startup. For testing purposes, the reset_database function can
be called before calling init_database again. This is useful to create a
fresh in-memory database for the test. Note that the Database objects
that were created in other modules at module-level scope will still be
valid and calling methods on them will be proxied to the new database
instance.

load_vms_tags() and load_vms_exists() have been modified so that they
are not called during import of web_utils but are called where needed
and the database is only contacted the first time--the resulting value
is cached. No database queries should be made simply as a result of
importing modules!