Added support for IMDS-based authentication to avoid using access/secret keys on EC2.
Using access/secret keys directly on EC2 is considered a bad security practice. I have added the option for boto3 to perform authentication using the Instance Metadata 
Service and assume the role associated with the EC2 instance on which CAPEv2 is running on.
This does not cancel the original authentication method; it is used as a fallback in case users do not specify secret/access keys in aws.conf.