Improve trend unquarantine processing time for non-quarantined samples.
Prior to these changes `trend_unquarantine` was passing the entire sample through a XOR routine before checking the XOR'd file header to determine if it was a quarantined file.

For a 100MB sample it was taking ~22 seconds to XOR the file.

```
%Own   %Total  OwnTime  TotalTime  Function
0.00%   0.00%   21.58s    21.58s   bytearray_xor (lib/cuckoo/common/quarantine.py)
```

By changing the function to only XOR the first 10 bytes of the file (header) before checking the file signature, we remove significant amounts of processing time for what will be almost all samples being submitted.

Using the same sample, post-change, the whole `unquarantine` process now only takes 0.770s.

```
%Own   %Total  OwnTime  TotalTime  Function
0.00%   0.00%   0.040s    0.770s   unquarantine (lib/cuckoo/common/quarantine.py)
```

Note, `unquarantine` is the parent function of `trend_unquarantine`, which calls `bytearray_xor`.