Azure machinery major updates (#2666)
* Force delete VMSS isntances for faster scaling down.

When instances are removed from the VMSS, we do not need to let them gracefully shut down.
Instances are using ephemeral OS disks, so no data persistence occurs. Thus, we don't care if the disk gets corrupted, each new instance starts fresh.

* Handle spot eviction IP conflict bug.

Spot instances can be removed form the VMSS and stick in our DB.
Azure should never be able to give two instances the same IP address, so it is safe to remove stale machine entry from DB.

* Update imports

* Fix az machinery logging typo

* Shift `find_machines_to_service_task` to abstract machinery class.

Allows individual machinery modules to modify task->machine search behavior.

* Add ability to run VMSS(s) with zero initial instances.

Removes the requirement to always have standing instances attached to VMSSs.
Uses fake "placeholder" machines in the DB to prevent crashing CAPE. Placeholders are removed during scale-ups and added when scaling to zero.

* Remove conditional for "per_platform" boolean assignment when calling `_thr_scale_machine_pool`

Value is required to be set in the configs or we bail out during initializing. Thus, can only ever be True in this type of conditional. Simplify the code to reflect that.

* Add docstring to abstract machinery function

* Remove unreachable code

* Update task-assignable VMSSs to filter on os_version and platform

* Reorder args for scale_from_zero method