analyzer/linux: Fix get_proc_status parsing (#2769)
* analyzer/linux: Fix get_proc_status parsing

When parsing the `/proc/<id>/status` file some field are empty
which made the dict() conversion fail because the list length was 1.

With this change, we are spliting on the ":" character and striping
the key and the value so we can get the key without the ":" at the end and
the value without any "\t" at the beginning or "\n" at the end.

* analyzer/linux: convert parent process id (PPid) to int